+++ src/logger.ts
@@ -50,9 +50,6 @@ export class Logger {
     }
 
     showLog() {
-        if (!this.extension.parser.buildLogRaw) {
-            return
-        }
         const uri = vscode.Uri.file(this.extension.manager.rootFile).with({scheme: 'latex-workshop-log'})
         let column = vscode.ViewColumn.Two
         if (vscode.window.activeTextEditor && vscode.window.activeTextEditor.viewColumn === vscode.ViewColumn.Two) {
@@ -65,9 +62,19 @@ export class Logger {
 
 export class LogProvider implements vscode.TextDocumentContentProvider {
     extension: Extension
+    change = new vscode.EventEmitter<vscode.Uri>()
 
     constructor(extension: Extension) {
         this.extension = extension
+    }
+
+    public update(uri: vscode.Uri) {
+        this.change.fire(uri)
+        console.log('change')
+    }
+
+    get onDidChange() : vscode.Event<vscode.Uri> {
+        return this.change.event
     }
 
     public provideTextDocumentContent(_uri: vscode.Uri) : string {
@@ -76,7 +83,6 @@ export class LogProvider implements vscode.TextDocumentContentProvider {
                                                                                          .replace(/>/g, "&gt;")
                                                                                          .replace(/"/g, "&quot;")
                                                                                          .replace(/'/g, "&#039;")}</span><br>`)
-        console.log(this.extension.parser.buildLogRaw)
         return `
             <!DOCTYPE html style="position:absolute; left: 0; top: 0; width: 100%; height: 100%;"><html><head></head>
             <body style="position:absolute; left: 0; top: 0; width: 100%; height: 100%; white-space: pre;">${dom.join('')}</body></html>